Amend how i'm using coordinates
Added new db object and now passing in % coordinated into db stored
proc. Struggling with unit tests as i dont know how to verify (or
Assert) values in a list.

Problem with populated a drop down as i have done for categories in
PitchUI as now i need params, which i dont have at the pitchUI level. So
instead of calling a method to "get" the drop down list, i need one that
returns to the mouse down event?